moved useClickOutside to hooks
