+++ packages/asc-ui/src/utils/hooks/index.ts
@@ -1,4 +1,5 @@
 export { default as useActionOnEscape } from './useActionOnEscape'
+export { default as useClickOutside } from './useClickOutside'
 export { default as useDebounce } from './useDebounce'
 export { default as useEdgeDetection } from './useEdgeDetection'
 export { default as useFocus } from './useFocus'
